@@ -1,14 +1,10 @@
 ext {
 
-    project.ext.compiledSdkVersion = 31
-    project.ext.kotlin_version = "1.5.31"
-    project.ext.gradle_build_tool_version = '7.0.4'
-
-    def androidx_lifecycle = "2.2.0"
-    def dagger_version = "2.36"
+    def kotlin_version = "1.8.10"
+    def androidx_lifecycle = "2.5.1"
     def android_arch_lifecycle_version = "1.1.1"
     def com_square_up_retrofit2_version = "2.4.0"
-    def room_version = "2.2.3"
+    def room_version = "2.5.1"
     def glide_version = "4.11.0"
     def rx_java_version = "2.2.6"
     def mockito_kotlin2_version = "3.2.0"
