@@ -1,24 +1,36 @@
 package com.haomins.data.db
 
+import android.content.Context
 import androidx.room.Database
+import androidx.room.Room
 import androidx.room.RoomDatabase
 import com.haomins.data.db.dao.ArticleDao
 import com.haomins.data.db.dao.SubscriptionDao
-import com.haomins.data.model.entities.ArticleEntity
-import com.haomins.data.model.entities.SubscriptionEntity
+import com.haomins.model.entity.ArticleEntity
+import com.haomins.model.entity.SubscriptionEntity
 
 @Database(
-        entities = [SubscriptionEntity::class, ArticleEntity::class],
-        version = 1,
-        exportSchema = false
+    entities = [SubscriptionEntity::class, ArticleEntity::class],
+    version = 1,
+    exportSchema = false
 )
 abstract class AppDatabase : RoomDatabase() {
-
-    companion object {
-        const val DATABASE_NAME = "READER_DB"
-    }
 
     abstract fun subscriptionDao(): SubscriptionDao
 
     abstract fun articleDao(): ArticleDao
+
+    companion object {
+
+        private const val DATABASE_NAME = "READER_DB"
+
+        fun getInstance(context: Context): AppDatabase {
+            return Room.databaseBuilder(
+                context,
+                AppDatabase::class.java,
+                DATABASE_NAME
+            ).build()
+        }
+
+    }
 }
