@@ -1,54 +1,54 @@
 package com.haomins.data
 
-import com.haomins.data.model.responses.article.*
-import com.haomins.data.model.responses.subscription.AddSourceResponseModel
-import com.haomins.data.model.responses.subscription.SubscriptionSourceListResponseModel
-import com.haomins.data.model.responses.user.UserAuthResponseModel
+import com.haomins.model.remote.article.*
+import com.haomins.model.remote.subscription.AddSourceResponseModel
+import com.haomins.model.remote.subscription.SubscriptionListResponseModel
+import com.haomins.model.remote.user.UserAuthResponseModel
 import com.haomins.data.service.TheOldReaderService
 import io.reactivex.Single
 
-class MockTheOldReaderService: TheOldReaderService {
+class MockTheOldReaderService : TheOldReaderService {
 
     override fun loginUser(
-            userEmail: String,
-            userPassword: String,
-            output: String,
-            service: String,
-            accountType: String,
-            client: String
+        userEmail: String,
+        userPassword: String,
+        output: String,
+        service: String,
+        accountType: String,
+        client: String
     ): Single<UserAuthResponseModel> {
         TODO("Not yet implemented")
     }
 
     override fun loadSubscriptionSourceList(
-            headerAuthValue: String,
-            output: String
-    ): Single<SubscriptionSourceListResponseModel> {
+        headerAuthValue: String,
+        output: String
+    ): Single<SubscriptionListResponseModel> {
         TODO("Not yet implemented")
     }
 
     override fun loadArticleListByFeed(
-            headerAuthValue: String,
-            feedId: String,
-            articleAmount: String,
-            continueLoad: String,
-            output: String
+        headerAuthValue: String,
+        feedId: String,
+        articleAmount: String,
+        continueLoad: String,
+        output: String
     ): Single<ItemRefListResponseModel> {
 
         fun mockReturn(): ItemRefListResponseModel {
             return ItemRefListResponseModel(
-                    itemRefs = mutableListOf<RefItemModel>().also {
-                        for (index in 0..9) {
-                            it.add(
-                                    RefItemModel(
-                                            "test_id",
-                                            mutableListOf("100$it", "100$it", "100$it"),
-                                            "100100100"
-                                    )
+                itemRefs = mutableListOf<RefItemModel>().also {
+                    for (index in 0..9) {
+                        it.add(
+                            RefItemModel(
+                                "test_id",
+                                mutableListOf("100$it", "100$it", "100$it"),
+                                "100100100"
                             )
-                        }
-                    },
-                    continuation = "test_continuation"
+                        )
+                    }
+                },
+                continuation = "test_continuation"
             )
         }
 
@@ -56,27 +56,27 @@ class MockTheOldReaderService: TheOldReaderService {
     }
 
     override fun loadAllArticles(
-            headerAuthValue: String,
-            articleAmount: String,
-            continueLoad: String,
-            allItemQuery: String,
-            output: String
+        headerAuthValue: String,
+        articleAmount: String,
+        continueLoad: String,
+        allItemQuery: String,
+        output: String
     ): Single<ItemRefListResponseModel> {
 
         fun mockReturn(): ItemRefListResponseModel {
             return ItemRefListResponseModel(
-                    itemRefs = mutableListOf<RefItemModel>().also {
-                        for (index in 0..9) {
-                            it.add(
-                                    RefItemModel(
-                                            id = "test_id_$index",
-                                            directStreamIds = mutableListOf("100$it", "100$it", "100$it"),
-                                            timestampUsec = "100100100"
-                                    )
+                itemRefs = mutableListOf<RefItemModel>().also {
+                    for (index in 0..9) {
+                        it.add(
+                            RefItemModel(
+                                id = "test_id_$index",
+                                directStreamIds = mutableListOf("100$it", "100$it", "100$it"),
+                                timestampUsec = "100100100"
                             )
-                        }
-                    },
-                    continuation = "test_continuation"
+                        )
+                    }
+                },
+                continuation = "test_continuation"
             )
         }
 
@@ -84,36 +84,36 @@ class MockTheOldReaderService: TheOldReaderService {
     }
 
     override fun loadArticleDetailsByRefId(
-            headerAuthValue: String,
-            refItemId: String,
-            output: String
+        headerAuthValue: String,
+        refItemId: String,
+        output: String
     ): Single<ArticleResponseModel> {
 
         fun mockReturn(): ArticleResponseModel {
             return ArticleResponseModel(
-                    direction = "test_direct",
-                    id = "test_id",
-                    title = "test_title",
-                    description = "test_description",
-                    updated = 0,
-                    items = mutableListOf<ArticleItemModel>().also {
-                        for (index in 0..9) {
-                            it.add(
-                                    ArticleItemModel(
-                                            crawlTimeMsec = "100100100",
-                                            timestampUsec = "100100100",
-                                            id = "test_id_$index",
-                                            categories = mutableListOf("cate_1", "cate_2", "cate_3"),
-                                            title = "test_title",
-                                            summary = Summary("test_direct", "test_content"),
-                                            publishedMillisecond = 1,
-                                            updatedMillisecond = 1,
-                                            author = "Tester Test",
-                                    )
+                direction = "test_direct",
+                id = "test_id",
+                title = "test_title",
+                description = "test_description",
+                updated = 0,
+                items = mutableListOf<ArticleItemModel>().also {
+                    for (index in 0..9) {
+                        it.add(
+                            ArticleItemModel(
+                                crawlTimeMsec = "100100100",
+                                timestampUsec = "100100100",
+                                id = "test_id_$index",
+                                categories = mutableListOf("cate_1", "cate_2", "cate_3"),
+                                title = "test_title",
+                                summary = Summary("test_direct", "test_content"),
+                                publishedMillisecond = 1,
+                                updatedMillisecond = 1,
+                                author = "Tester Test",
                             )
-                        }
-                    },
-                    likingUsersCount = 1,
+                        )
+                    }
+                },
+                likingUsersCount = 1,
                 alternate = Alternate("www.test.com", "www.test.com/test.jpeg")
             )
         }
@@ -122,9 +122,9 @@ class MockTheOldReaderService: TheOldReaderService {
     }
 
     override fun addSubscription(
-            headerAuthValue: String,
-            quickAddSubscription: String,
-            output: String
+        headerAuthValue: String,
+        quickAddSubscription: String,
+        output: String
     ): Single<AddSourceResponseModel> {
         TODO("Not yet implemented")
     }
