@@ -1,3 +1,4 @@
 package com.haomins.domain.exception
 
-class DuplicateRequestException : IllegalStateException("The Use Case Can Only Be Executed One At a Time")
+class DuplicateRequestException :
+    IllegalStateException("The Use Case Can Only Be Executed One At a Time")
