@@ -1,3 +1,4 @@
 package com.haomins.domain.exception
 
-class ParamsShouldNotBeNullException : IllegalArgumentException("The Use Case Params Should Not Be Null")
+class ParamsShouldNotBeNullException :
+    IllegalArgumentException("The Use Case Params Should Not Be Null")
