@@ -3,16 +3,21 @@ package com.haomins.reader.view.activities
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.haomins.reader.R
+import com.haomins.reader.databinding.ActivityAddSourceBinding
 import com.haomins.reader.utils.slideInAnimation
 import com.haomins.reader.utils.slideOutAnimation
 import com.haomins.reader.view.fragments.AddSourceFragment
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class AddSourceActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityAddSourceBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        inflateView()
         slideInAnimation()
-        setContentView(R.layout.activity_add_source)
         showAddSourceFragment()
     }
 
@@ -21,11 +26,16 @@ class AddSourceActivity : AppCompatActivity() {
         slideOutAnimation()
     }
 
+    private fun inflateView() {
+        binding = ActivityAddSourceBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+
     private fun showAddSourceFragment() {
         supportFragmentManager.beginTransaction().replace(
-                R.id.add_source_activity_frame_layout,
-                AddSourceFragment(),
-                AddSourceFragment.TAG
+            R.id.add_source_activity_frame_layout,
+            AddSourceFragment(),
+            AddSourceFragment.TAG
         ).commit()
     }
 }
