@@ -1,25 +1,39 @@
 package com.haomins.reader.view.activities
 
 import android.os.Bundle
+import android.util.Log
+import android.view.KeyEvent
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.haomins.reader.R
+import androidx.viewpager2.widget.ViewPager2.OnPageChangeCallback
+import com.haomins.reader.databinding.ActivityArticleDetailBinding
 import com.haomins.reader.utils.slideInAnimation
 import com.haomins.reader.utils.slideOutAnimation
-import com.haomins.reader.view.fragments.ArticleDetailFragment
-import kotlinx.android.synthetic.main.activity_article_detail.*
+import com.haomins.reader.view.fragments.articles.ArticleDetailFragment
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class ArticleDetailActivity : AppCompatActivity() {
 
     companion object {
         const val ARTICLE_ITEM_ID = "ARTICLE_ITEM_ID"
+        private const val TAG = "ArticleDetailActivity"
     }
+
+    private var currentPosition: Int = -1
+    private val articleIdArray by lazy {
+        intent.getStringArrayExtra(ArticleListActivity.ARTICLE_ITEM_ID_ARRAY)
+    }
+    private val articleId by lazy {
+        intent.getStringExtra(ArticleListActivity.ARTICLE_ITEM_ID)
+    }
+    private lateinit var binding: ActivityArticleDetailBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        inflateView()
         slideInAnimation()
-        setContentView(R.layout.activity_article_detail)
         initViewPager()
     }
 
@@ -28,18 +42,50 @@ class ArticleDetailActivity : AppCompatActivity() {
         slideOutAnimation()
     }
 
+    override fun onKeyDown(keyCode: Int, event: KeyEvent?): Boolean {
+        Log.d(TAG, "onKeyDown :: keyCode is $keyCode, view pager position is $currentPosition")
+        when (keyCode) {
+            KeyEvent.KEYCODE_VOLUME_DOWN -> {
+                currentPosition += 1
+                binding.articleDetailViewPager.setCurrentItem(currentPosition, true)
+            }
+            KeyEvent.KEYCODE_VOLUME_UP -> {
+                if (currentPosition > 0) {
+                    currentPosition -= 1
+                    binding.articleDetailViewPager.setCurrentItem(currentPosition, true)
+                }
+            }
+            else -> {
+                super.onKeyDown(keyCode, event)
+            }
+        }
+        return true
+    }
+
     private fun initViewPager() {
-        val articleId = intent.getStringExtra(ArticleListActivity.ARTICLE_ITEM_ID)
-        val articleIdArray = intent.getStringArrayExtra(ArticleListActivity.ARTICLE_ITEM_ID_ARRAY)
         articleIdArray?.let {
-            val currentPosition = articleIdArray.indexOf(articleId)
+            currentPosition = it.indexOf(articleId)
             val adapter = ArticleDetailFragmentAdapter(
                 this,
                 it
             )
-            article_detail_view_pager.adapter = adapter
-            article_detail_view_pager.setCurrentItem(currentPosition, false)
+            binding.articleDetailViewPager.adapter = adapter
+            binding.articleDetailViewPager.setCurrentItem(currentPosition, false)
+            binding.articleDetailViewPager.registerOnPageChangeCallback(
+                object : OnPageChangeCallback() {
+                    override fun onPageSelected(position: Int) {
+                        currentPosition = position
+                        Log.d(TAG, "current position is $currentPosition")
+                        super.onPageSelected(position)
+                    }
+                }
+            )
         }
+    }
+
+    private fun inflateView() {
+        binding = ActivityArticleDetailBinding.inflate(layoutInflater)
+        setContentView(binding.root)
     }
 
     private inner class ArticleDetailFragmentAdapter(
