@@ -1,6 +1,5 @@
 package com.haomins.reader.view.fragments
 
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,21 +9,19 @@ import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import com.haomins.reader.BuildConfig
 import com.haomins.reader.R
-import com.haomins.reader.ReaderApplication
+import com.haomins.reader.databinding.FragmentLoginBinding
 import com.haomins.reader.utils.showToast
 import com.haomins.reader.view.activities.MainActivity
 import com.haomins.reader.viewModels.LoginViewModel
-import kotlinx.android.synthetic.main.fragment_login.*
-import javax.inject.Inject
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class LoginFragment : Fragment() {
 
-    @Inject
-    lateinit var viewModelFactory: ViewModelProvider.Factory
-    private val loginViewModel by viewModels<LoginViewModel> { viewModelFactory }
+    private val loginViewModel by viewModels<LoginViewModel>()
+    private lateinit var binding: FragmentLoginBinding
 
     private val isUserLoggedInObserver by lazy {
         Observer<Boolean> {
@@ -35,18 +32,13 @@ class LoginFragment : Fragment() {
         }
     }
 
-    override fun onAttach(context: Context) {
-        (requireActivity().application as ReaderApplication).appComponent.viewModelComponent()
-            .build().inject(this)
-        super.onAttach(context)
-    }
-
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_login, container, false)
+    ): View {
+        binding = FragmentLoginBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -58,10 +50,12 @@ class LoginFragment : Fragment() {
     }
 
     private fun setOnClickListener() {
-        login_button.setOnClickListener { loginButtonOnClick() }
-        sign_up_button.setOnClickListener { signUpButtonOnClick() }
-        forgot_password.setOnClickListener { forgetPasswordButtonOnClick() }
-        news_app_disclosure.setOnClickListener { showDisclosureFragment() }
+        with(binding) {
+            loginButton.setOnClickListener { loginButtonOnClick() }
+            signUpButton.setOnClickListener { signUpButtonOnClick() }
+            forgotPassword.setOnClickListener { forgetPasswordButtonOnClick() }
+            newsAppDisclosure.setOnClickListener { showDisclosureFragment() }
+        }
     }
 
     private fun showDisclosureFragment() {
@@ -73,13 +67,15 @@ class LoginFragment : Fragment() {
     }
 
     private fun setLoginPasswordEditTextOnChangeListener() {
-        login_password_edit_text.addTextChangedListener {
-            login_button.isEnabled = !it.isNullOrEmpty()
+        with(binding) {
+            loginPasswordEditText.addTextChangedListener {
+                loginButton.isEnabled = !it.isNullOrEmpty()
+            }
         }
     }
 
     private fun initiateUI() {
-        login_app_version_text_view.text =
+        binding.loginAppVersionTextView.text =
             getString(R.string.version_description, BuildConfig.VERSION_NAME)
     }
 
@@ -100,11 +96,13 @@ class LoginFragment : Fragment() {
     }
 
     private fun loginButtonOnClick() {
-        loginViewModel.authorize(
-            userName = login_username_edit_text.text.toString(),
-            userPassword = login_password_edit_text.text.toString()
-        ) {
-            showToast(it)
+        with(binding) {
+            loginViewModel.authorize(
+                userName = loginUsernameEditText.text.toString(),
+                userPassword = loginPasswordEditText.text.toString()
+            ) {
+                showToast(it)
+            }
         }
     }
 
