@@ -3,12 +3,14 @@ package com.haomins.reader.viewModels
 import android.app.Application
 import android.util.Log
 import androidx.lifecycle.ViewModel
-import com.haomins.domain.model.DisclosureInfo
 import com.haomins.domain.usecase.disclosure.LoadDisclosureContent
+import com.haomins.model.DisclosureInfo
 import com.haomins.reader.R
+import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.observers.DisposableSingleObserver
 import javax.inject.Inject
 
+@HiltViewModel
 class DisclosureViewModel @Inject constructor(
     private val application: Application,
     private val loadDisclosureContent: LoadDisclosureContent
